Changed how VCBacklog sorts the backlog
This commit moves the sorting logic from BacklogView to a separate Sorting Decorator for the Backlog. Sorting is not a responsibility of the View logic. It's also easier to change the sorting later this way. The commit also adds an Interface for the View, but that's just a minor addition.